+++ django/applications/catmaid/control/project.py
@@ -7,13 +7,15 @@ import yaml
 from guardian.shortcuts import get_objects_for_user
 
 from django.db import connection
+from django.db.models import Exists, OuterRef
 from django.http import HttpRequest, HttpResponse, JsonResponse
 from django.shortcuts import get_object_or_404
 
 from catmaid.models import (BrokenSlice, Class, ClientDatastore,
-        InterpolatableSection, Project, ProjectStack, Relation, Stack,
+        InterpolatableSection, Location, Project, ProjectStack, Relation, Stack,
         StackGroup, StackStackGroup, UserRole)
 from catmaid.control.authentication import requires_user_role
+from catmaid.control.common import get_request_bool
 
 from rest_framework.decorators import api_view
 
@@ -183,6 +185,12 @@ def projects(request:HttpRequest) -> JsonResponse:
             items:
               $ref: project_api_stackgroup_element
             required: true
+    parameters:
+    - name: has_tracing_data
+      description: Return only projects that have tracing data
+      required: false
+      defaultValue: falese
+      type: boolean
     type:
       projects:
         type: array
@@ -193,6 +201,14 @@ def projects(request:HttpRequest) -> JsonResponse:
 
     # Get all projects that are visisble for the current user
     projects = get_project_qs_for_user(request.user).order_by('title')
+    has_tracing_data = get_request_bool(request.GET, 'has_tracing_data', False)
+
+    if has_tracing_data:
+        projects = projects.annotate(
+                no_locations=~Exists(Location.objects.filter(project=OuterRef('pk')))
+            ).filter(
+                no_locations=False
+            )
 
     if 0 == len(projects):
         return JsonResponse([], safe=False)
@@ -200,6 +216,19 @@ def projects(request:HttpRequest) -> JsonResponse:
     cursor = connection.cursor()
     project_template = ",".join(("(%s)",) * len(projects)) or "()"
     user_project_ids = [p.id for p in projects]
+
+    tracing_data_join = ''
+    extra_where = []
+    if has_tracing_data:
+        tracing_data_join = '''
+            INNER JOIN LATERAL (
+                SELECT EXISTS (SELECT 1 FROM location WHERE project_id = ps.project_id)
+            ) sub(has_tracing_data)
+                ON TRUE
+        '''
+        extra_where.append('''
+            sub.has_tracing_data = True
+        ''')
 
     cursor.execute(f"""
         SELECT ps.project_id, ps.stack_id, s.title, s.comment FROM project_stack ps