+++ django/applications/catmaid/static/js/widgets/overlay.js
@@ -3,7 +3,6 @@
 /* global
   CATMAID,
   project,
-  user_groups,
   msgpack
 */